Temporarily move fixture into test file
